Creating two pathways (speed vs consistency) + storing embeddings as doubles

+++ cpp/include/cuml/manifold/umap.hpp
@@ -23,19 +23,19 @@
 namespace ML {
 
 void transform(const cumlHandle &handle, float *X, int n, int d, float *orig_X,
-               int orig_n, float *embedding, int embedding_n,
-               UMAPParams *params, float *transformed);
+               int orig_n, double *embedding, int embedding_n,
+               UMAPParams *params, double *transformed);
 
 void fit(const cumlHandle &handle,
          float *X,  // input matrix
          float *y,  // labels
-         int n, int d, UMAPParams *params, float *embeddings);
+         int n, int d, UMAPParams *params, double *embeddings);
 
 void fit(const cumlHandle &handle,
          float *X,  // input matrix
          int n,     // rows
          int d,     // cols
-         UMAPParams *params, float *embeddings);
+         UMAPParams *params, double *embeddings);
 
 class UMAP_API {
   float *orig_X;
@@ -58,7 +58,7 @@ class UMAP_API {
    * @param embeddings
    *        an array to return the output embeddings of size (n_samples, n_components)
    */
-  void fit(float *X, int n, int d, float *embeddings);
+  void fit(float *X, int n, int d, double *embeddings);
 
   /**
    * Fits a supervised UMAP model
@@ -73,7 +73,7 @@ class UMAP_API {
    * @param embeddings
    *        an array to return the output embeddings of size (n_samples, n_components)
    */
-  void fit(float *X, float *y, int n, int d, float *embeddings);
+  void fit(float *X, float *y, int n, int d, double *embeddings);
 
   /**
    * Project a set of X vectors into the embedding space.
@@ -90,8 +90,8 @@ class UMAP_API {
    * @param out
    *        pointer to array for storing output embeddings (n, n_components)
    */
-  void transform(float *X, int n, int d, float *embedding, int embedding_n,
-                 float *out);
+  void transform(float *X, int n, int d, double *embedding, int embedding_n,
+                 double *out);
 
   /**
    * Get the UMAPParams instance
