+++ cpp/src/umap/umap.cu
@@ -25,15 +25,15 @@ namespace ML {
 static const int TPB_X = 256;
 
 void transform(const cumlHandle &handle, float *X, int n, int d, float *orig_X,
-               int orig_n, float *embedding, int embedding_n,
-               UMAPParams *params, float *transformed) {
+               int orig_n, double *embedding, int embedding_n,
+               UMAPParams *params, double *transformed) {
   UMAPAlgo::_transform<float, TPB_X>(handle, X, n, d, orig_X, orig_n, embedding,
                                      embedding_n, params, transformed);
 }
 void fit(const cumlHandle &handle,
          float *X,  // input matrix
          float *y,  // labels
-         int n, int d, UMAPParams *params, float *embeddings) {
+         int n, int d, UMAPParams *params, double *embeddings) {
   UMAPAlgo::_fit<float, TPB_X>(handle, X, y, n, d, params, embeddings);
 }
 
@@ -41,7 +41,7 @@ void fit(const cumlHandle &handle,
          float *X,  // input matrix
          int n,     // rows
          int d,     // cols
-         UMAPParams *params, float *embeddings) {
+         UMAPParams *params, double *embeddings) {
   UMAPAlgo::_fit<float, TPB_X>(handle, X, n, d, params, embeddings);
 }
 UMAP_API::UMAP_API(const cumlHandle &handle, UMAPParams *params)
@@ -64,14 +64,14 @@ UMAP_API::~UMAP_API() {}
  * @param embeddings
  *        an array to return the output embeddings of size (n_samples, n_components)
  */
-void UMAP_API::fit(float *X, int n, int d, float *embeddings) {
+void UMAP_API::fit(float *X, int n, int d, double *embeddings) {
   this->orig_X = X;
   this->orig_n = n;
   UMAPAlgo::_fit<float, TPB_X>(*this->handle, X, n, d, get_params(),
                                embeddings);
 }
 
-void UMAP_API::fit(float *X, float *y, int n, int d, float *embeddings) {
+void UMAP_API::fit(float *X, float *y, int n, int d, double *embeddings) {
   this->orig_X = X;
   this->orig_n = n;
 
@@ -94,8 +94,8 @@ void UMAP_API::fit(float *X, float *y, int n, int d, float *embeddings) {
  * @param out
  *        pointer to array for storing output embeddings (n, n_components)
  */
-void UMAP_API::transform(float *X, int n, int d, float *embedding,
-                         int embedding_n, float *out) {
+void UMAP_API::transform(float *X, int n, int d, double *embedding,
+                         int embedding_n, double *out) {
   UMAPAlgo::_transform<float, TPB_X>(*this->handle, X, n, d, this->orig_X,
                                      this->orig_n, embedding, embedding_n,
                                      get_params(), out);
